Include optional dependency - [VTK]
OK, the optional dependency vtk is now included in pyproject.toml. Tested with pip, pipx and conda, with and without vtk previously installed... seems to work fine on my machine.

Updated the readme to reflect installation with the optional dependency.

Updated common.py to have an legible Exception message when vtk is not available.